requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.1.10
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
Faker==19.3.1
idna==3.4
iniconfig==2.0.0
Markdown==3.4.4
packaging==23.2
pluggy==1.3.0
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
